qcacmn: Add structural, API prototype and definition changes
The following changes are made as part of this commit:

1] Create new arrays 'chan_psd_power_info' and 'chan_eirp_power_info'
   in structure 'reg_tpc_power_info' for storing PSD and EIRP power
   values respectively. Also, create the corresponding array size
   variables. These new arrays and variables should be filled only
   for SP power type based on the WMI service advertised by HALphy
   indicating support for both PSD and EIRP power values.
2] Change the power type from unsigned integer to signed integer in
   'reg_tpc_power_info' as negative power values are valid. Use
   16-bit integer for storing power values wherever possible.
3] Change the relevant APIs' prototype declaration and
   definition to use signed integer instead of unsigned values.
4] In API, 'reg_get_sp_eirp', EIRP power (both AP and Client) for
   punctured channels is calculated as
   MIN(EIRP_FROM_OOBE_PSD, AP_REG_EIRP). This calculation is correct
   for AP power types, but for client power, we should limit the
   AFC EIRP by CLIENT_REG_EIRP. Fix this by selecting the regulatory
   EIRP from appropriate master channel list based on AP / Client
   types.

CRs-Fixed: 3692772

+++ umac/regulatory/core/src/reg_services_common.h
@@ -2089,8 +2089,8 @@ QDF_STATUS reg_get_client_power_for_connecting_ap(struct wlan_objmgr_pdev *pdev,
 						  enum reg_6g_ap_type ap_type,
 						  qdf_freq_t chan_freq,
 						  bool is_psd,
-						  uint16_t *tx_power,
-						  uint16_t *eirp_psd_power);
+						  int16_t *tx_power,
+						  int16_t *eirp_psd_power);
 
 /**
  * reg_get_client_power_for_6ghz_ap() - Find the channel information when
@@ -2113,8 +2113,8 @@ QDF_STATUS reg_get_client_power_for_connecting_ap(struct wlan_objmgr_pdev *pdev,
 QDF_STATUS reg_get_client_power_for_6ghz_ap(struct wlan_objmgr_pdev *pdev,
 					    enum reg_6g_client_type client_type,
 					    qdf_freq_t chan_freq,
-					    bool *is_psd, uint16_t *tx_power,
-					    uint16_t *eirp_psd_power);
+					    bool *is_psd, int16_t *tx_power,
+					    int16_t *eirp_psd_power);
 
 /**
  * reg_set_ap_pwr_and_update_chan_list() - Set the AP power mode and recompute
@@ -2233,8 +2233,8 @@ QDF_STATUS reg_get_client_power_for_connecting_ap(struct wlan_objmgr_pdev *pdev,
 						  enum reg_6g_ap_type ap_type,
 						  qdf_freq_t chan_freq,
 						  bool is_psd,
-						  uint16_t *tx_power,
-						  uint16_t *eirp_psd_power)
+						  int16_t *tx_power,
+						  int16_t *eirp_psd_power)
 {
 	*tx_power = 0;
 	*eirp_psd_power = 0;
@@ -2245,8 +2245,8 @@ static inline
 QDF_STATUS reg_get_client_power_for_6ghz_ap(struct wlan_objmgr_pdev *pdev,
 					    enum reg_6g_client_type client_type,
 					    qdf_freq_t chan_freq,
-					    bool *is_psd, uint16_t *tx_power,
-					    uint16_t *eirp_psd_power)
+					    bool *is_psd, int16_t *tx_power,
+					    int16_t *eirp_psd_power)
 {
 	*is_psd = false;
 	*tx_power = 0;
@@ -2776,12 +2776,12 @@ enum reg_6g_ap_type reg_get_best_pwr_mode(struct wlan_objmgr_pdev *pdev,
  *
  * Return: EIRP power
  */
-int8_t reg_get_eirp_pwr(struct wlan_objmgr_pdev *pdev, qdf_freq_t freq,
-			qdf_freq_t cen320,
-			uint16_t bw, enum reg_6g_ap_type ap_pwr_type,
-			uint16_t in_punc_pattern,
-			bool is_client_list_lookup_needed,
-			enum reg_6g_client_type client_type);
+int16_t reg_get_eirp_pwr(struct wlan_objmgr_pdev *pdev, qdf_freq_t freq,
+			 qdf_freq_t cen320,
+			 uint16_t bw, enum reg_6g_ap_type ap_pwr_type,
+			 uint16_t in_punc_pattern,
+			 bool is_client_list_lookup_needed,
+			 enum reg_6g_client_type client_type);
 
 #else
 static inline
